combined 2 client files and 2 server files into 1 server and 1 client file
